+++ dnacentersdk/api/__init__.py
@@ -331,82 +331,82 @@ from .v2_3_5_3.users import \
 from .v2_3_5_3.wireless import \
     Wireless as Wireless_v2_3_5_3
 
-from .v2_3_7_6.application_policy import \
-    ApplicationPolicy as ApplicationPolicy_v2_3_7_6
-from .v2_3_7_6.applications import \
-    Applications as Applications_v2_3_7_6
-from .v2_3_7_6.clients import \
-    Clients as Clients_v2_3_7_6
-from .v2_3_7_6.command_runner import \
-    CommandRunner as CommandRunner_v2_3_7_6
-from .v2_3_7_6.compliance import \
-    Compliance as Compliance_v2_3_7_6
-from .v2_3_7_6.configuration_archive import \
-    ConfigurationArchive as ConfigurationArchive_v2_3_7_6
-from .v2_3_7_6.configuration_templates import \
-    ConfigurationTemplates as ConfigurationTemplates_v2_3_7_6
-from .v2_3_7_6.device_onboarding_pnp import \
-    DeviceOnboardingPnp as DeviceOnboardingPnp_v2_3_7_6
-from .v2_3_7_6.device_replacement import \
-    DeviceReplacement as DeviceReplacement_v2_3_7_6
-from .v2_3_7_6.devices import \
-    Devices as Devices_v2_3_7_6
-from .v2_3_7_6.discovery import \
-    Discovery as Discovery_v2_3_7_6
-from .v2_3_7_6.eox import \
-    EoX as EoX_v2_3_7_6
-from .v2_3_7_6.event_management import \
-    EventManagement as EventManagement_v2_3_7_6
-from .v2_3_7_6.fabric_wireless import \
-    FabricWireless as FabricWireless_v2_3_7_6
-from .v2_3_7_6.file import \
-    File as File_v2_3_7_6
-from .v2_3_7_6.health_and_performance import \
-    HealthAndPerformance as HealthAndPerformance_v2_3_7_6
-from .v2_3_7_6.itsm import \
-    Itsm as Itsm_v2_3_7_6
-from .v2_3_7_6.itsm_integration import \
-    ItsmIntegration as ItsmIntegration_v2_3_7_6
-from .v2_3_7_6.issues import \
-    Issues as Issues_v2_3_7_6
-from .v2_3_7_6.lan_automation import \
-    LanAutomation as LanAutomation_v2_3_7_6
-from .v2_3_7_6.licenses import \
-    Licenses as Licenses_v2_3_7_6
-from .v2_3_7_6.network_settings import \
-    NetworkSettings as NetworkSettings_v2_3_7_6
-from .v2_3_7_6.path_trace import \
-    PathTrace as PathTrace_v2_3_7_6
-from .v2_3_7_6.platform import \
-    Platform as Platform_v2_3_7_6
-from .v2_3_7_6.reports import \
-    Reports as Reports_v2_3_7_6
-from .v2_3_7_6.sda import \
-    Sda as Sda_v2_3_7_6
-from .v2_3_7_6.security_advisories import \
-    SecurityAdvisories as SecurityAdvisories_v2_3_7_6
-from .v2_3_7_6.sensors import \
-    Sensors as Sensors_v2_3_7_6
-from .v2_3_7_6.site_design import \
-    SiteDesign as SiteDesign_v2_3_7_6
-from .v2_3_7_6.sites import \
-    Sites as Sites_v2_3_7_6
-from .v2_3_7_6.software_image_management_swim import \
-    SoftwareImageManagementSwim as SoftwareImageManagementSwim_v2_3_7_6
-from .v2_3_7_6.system_settings import \
-    SystemSettings as SystemSettings_v2_3_7_6
-from .v2_3_7_6.tag import \
-    Tag as Tag_v2_3_7_6
-from .v2_3_7_6.task import \
-    Task as Task_v2_3_7_6
-from .v2_3_7_6.topology import \
-    Topology as Topology_v2_3_7_6
-from .v2_3_7_6.user_and_roles import \
-    UserandRoles as UserandRoles_v2_3_7_6
-from .v2_3_7_6.users import \
-    Users as Users_v2_3_7_6
-from .v2_3_7_6.wireless import \
-    Wireless as Wireless_v2_3_7_6
+from .v2_3_7_5.application_policy import \
+    ApplicationPolicy as ApplicationPolicy_v2_3_7_5
+from .v2_3_7_5.applications import \
+    Applications as Applications_v2_3_7_5
+from .v2_3_7_5.clients import \
+    Clients as Clients_v2_3_7_5
+from .v2_3_7_5.command_runner import \
+    CommandRunner as CommandRunner_v2_3_7_5
+from .v2_3_7_5.compliance import \
+    Compliance as Compliance_v2_3_7_5
+from .v2_3_7_5.configuration_archive import \
+    ConfigurationArchive as ConfigurationArchive_v2_3_7_5
+from .v2_3_7_5.configuration_templates import \
+    ConfigurationTemplates as ConfigurationTemplates_v2_3_7_5
+from .v2_3_7_5.device_onboarding_pnp import \
+    DeviceOnboardingPnp as DeviceOnboardingPnp_v2_3_7_5
+from .v2_3_7_5.device_replacement import \
+    DeviceReplacement as DeviceReplacement_v2_3_7_5
+from .v2_3_7_5.devices import \
+    Devices as Devices_v2_3_7_5
+from .v2_3_7_5.discovery import \
+    Discovery as Discovery_v2_3_7_5
+from .v2_3_7_5.eox import \
+    EoX as EoX_v2_3_7_5
+from .v2_3_7_5.event_management import \
+    EventManagement as EventManagement_v2_3_7_5
+from .v2_3_7_5.fabric_wireless import \
+    FabricWireless as FabricWireless_v2_3_7_5
+from .v2_3_7_5.file import \
+    File as File_v2_3_7_5
+from .v2_3_7_5.health_and_performance import \
+    HealthAndPerformance as HealthAndPerformance_v2_3_7_5
+from .v2_3_7_5.itsm import \
+    Itsm as Itsm_v2_3_7_5
+from .v2_3_7_5.itsm_integration import \
+    ItsmIntegration as ItsmIntegration_v2_3_7_5
+from .v2_3_7_5.issues import \
+    Issues as Issues_v2_3_7_5
+from .v2_3_7_5.lan_automation import \
+    LanAutomation as LanAutomation_v2_3_7_5
+from .v2_3_7_5.licenses import \
+    Licenses as Licenses_v2_3_7_5
+from .v2_3_7_5.network_settings import \
+    NetworkSettings as NetworkSettings_v2_3_7_5
+from .v2_3_7_5.path_trace import \
+    PathTrace as PathTrace_v2_3_7_5
+from .v2_3_7_5.platform import \
+    Platform as Platform_v2_3_7_5
+from .v2_3_7_5.reports import \
+    Reports as Reports_v2_3_7_5
+from .v2_3_7_5.sda import \
+    Sda as Sda_v2_3_7_5
+from .v2_3_7_5.security_advisories import \
+    SecurityAdvisories as SecurityAdvisories_v2_3_7_5
+from .v2_3_7_5.sensors import \
+    Sensors as Sensors_v2_3_7_5
+from .v2_3_7_5.site_design import \
+    SiteDesign as SiteDesign_v2_3_7_5
+from .v2_3_7_5.sites import \
+    Sites as Sites_v2_3_7_5
+from .v2_3_7_5.software_image_management_swim import \
+    SoftwareImageManagementSwim as SoftwareImageManagementSwim_v2_3_7_5
+from .v2_3_7_5.system_settings import \
+    SystemSettings as SystemSettings_v2_3_7_5
+from .v2_3_7_5.tag import \
+    Tag as Tag_v2_3_7_5
+from .v2_3_7_5.task import \
+    Task as Task_v2_3_7_5
+from .v2_3_7_5.topology import \
+    Topology as Topology_v2_3_7_5
+from .v2_3_7_5.user_and_roles import \
+    UserandRoles as UserandRoles_v2_3_7_5
+from .v2_3_7_5.users import \
+    Users as Users_v2_3_7_5
+from .v2_3_7_5.wireless import \
+    Wireless as Wireless_v2_3_7_5
 from .custom_caller import CustomCaller
 
 
@@ -503,7 +503,7 @@ class DNACenterAPI(object):
                 argument or an environment variable, or it is not a
                 DNA Center API supported version
                 [  '2.2.2.3', '2.2.3.3',
-                '2.3.3.0', '2.3.5.3', '2.3.7.6'].
+                '2.3.3.0', '2.3.5.3', '2.3.7.5'].
 
         """
         username = username or dnacenter_environment.get_env_username()
@@ -537,12 +537,12 @@ class DNACenterAPI(object):
 
         if version not in ['2.2.2.3', '2.2.3.3',
                            '2.3.3.0', '2.3.5.3',
-                           '2.3.7.6']:
+                           '2.3.7.5']:
             raise VersionError(
                 'Unknown API version, '
                 + 'known versions are {}'.format(
                     '  2.2.2.3, 2.2.3.3, 2.3.3.0, 2.3.5.3 and '
-                    + '2.3.7.6'
+                    + '2.3.7.5'
                 )
             )
 
@@ -1181,157 +1181,157 @@ class DNACenterAPI(object):
                     self._session, object_factory, _validator
                 )
 
-        if version == '2.3.7.6':
+        if version == '2.3.7.5':
             self.application_policy = \
-                ApplicationPolicy_v2_3_7_6(
+                ApplicationPolicy_v2_3_7_5(
                     self._session, object_factory, _validator
                 )
             self.applications = \
-                Applications_v2_3_7_6(
+                Applications_v2_3_7_5(
                     self._session, object_factory, _validator
                 )
             self.clients = \
-                Clients_v2_3_7_6(
+                Clients_v2_3_7_5(
                     self._session, object_factory, _validator
                 )
             self.command_runner = \
-                CommandRunner_v2_3_7_6(
+                CommandRunner_v2_3_7_5(
                     self._session, object_factory, _validator
                 )
             self.compliance = \
-                Compliance_v2_3_7_6(
+                Compliance_v2_3_7_5(
                     self._session, object_factory, _validator
                 )
             self.configuration_archive = \
-                ConfigurationArchive_v2_3_7_6(
+                ConfigurationArchive_v2_3_7_5(
                     self._session, object_factory, _validator
                 )
             self.configuration_templates = \
-                ConfigurationTemplates_v2_3_7_6(
+                ConfigurationTemplates_v2_3_7_5(
                     self._session, object_factory, _validator
                 )
             self.device_onboarding_pnp = \
-                DeviceOnboardingPnp_v2_3_7_6(
+                DeviceOnboardingPnp_v2_3_7_5(
                     self._session, object_factory, _validator
                 )
             self.device_replacement = \
-                DeviceReplacement_v2_3_7_6(
+                DeviceReplacement_v2_3_7_5(
                     self._session, object_factory, _validator
                 )
             self.devices = \
-                Devices_v2_3_7_6(
+                Devices_v2_3_7_5(
                     self._session, object_factory, _validator
                 )
             self.discovery = \
-                Discovery_v2_3_7_6(
+                Discovery_v2_3_7_5(
                     self._session, object_factory, _validator
                 )
             self.eox = \
-                EoX_v2_3_7_6(
+                EoX_v2_3_7_5(
                     self._session, object_factory, _validator
                 )
             self.event_management = \
-                EventManagement_v2_3_7_6(
+                EventManagement_v2_3_7_5(
                     self._session, object_factory, _validator
                 )
             self.fabric_wireless = \
-                FabricWireless_v2_3_7_6(
+                FabricWireless_v2_3_7_5(
                     self._session, object_factory, _validator
                 )
             self.file = \
-                File_v2_3_7_6(
+                File_v2_3_7_5(
                     self._session, object_factory, _validator
                 )
             self.health_and_performance = \
-                HealthAndPerformance_v2_3_7_6(
+                HealthAndPerformance_v2_3_7_5(
                     self._session, object_factory, _validator
                 )
             self.itsm = \
-                Itsm_v2_3_7_6(
+                Itsm_v2_3_7_5(
                     self._session, object_factory, _validator
                 )
             self.itsm_integration = \
-                ItsmIntegration_v2_3_7_6(
+                ItsmIntegration_v2_3_7_5(
                     self._session, object_factory, _validator
                 )
             self.issues = \
-                Issues_v2_3_7_6(
+                Issues_v2_3_7_5(
                     self._session, object_factory, _validator
                 )
             self.lan_automation = \
-                LanAutomation_v2_3_7_6(
+                LanAutomation_v2_3_7_5(
                     self._session, object_factory, _validator
                 )
             self.licenses = \
-                Licenses_v2_3_7_6(
+                Licenses_v2_3_7_5(
                     self._session, object_factory, _validator
                 )
             self.network_settings = \
-                NetworkSettings_v2_3_7_6(
+                NetworkSettings_v2_3_7_5(
                     self._session, object_factory, _validator
                 )
             self.path_trace = \
-                PathTrace_v2_3_7_6(
+                PathTrace_v2_3_7_5(
                     self._session, object_factory, _validator
                 )
             self.platform = \
-                Platform_v2_3_7_6(
+                Platform_v2_3_7_5(
                     self._session, object_factory, _validator
                 )
             self.reports = \
-                Reports_v2_3_7_6(
+                Reports_v2_3_7_5(
                     self._session, object_factory, _validator
                 )
             self.sda = \
-                Sda_v2_3_7_6(
+                Sda_v2_3_7_5(
                     self._session, object_factory, _validator
                 )
             self.security_advisories = \
-                SecurityAdvisories_v2_3_7_6(
+                SecurityAdvisories_v2_3_7_5(
                     self._session, object_factory, _validator
                 )
             self.sensors = \
-                Sensors_v2_3_7_6(
+                Sensors_v2_3_7_5(
                     self._session, object_factory, _validator
                 )
             self.site_design = \
-                SiteDesign_v2_3_7_6(
+                SiteDesign_v2_3_7_5(
                     self._session, object_factory, _validator
                 )
             self.sites = \
-                Sites_v2_3_7_6(
+                Sites_v2_3_7_5(
                     self._session, object_factory, _validator
                 )
             self.software_image_management_swim = \
-                SoftwareImageManagementSwim_v2_3_7_6(
+                SoftwareImageManagementSwim_v2_3_7_5(
                     self._session, object_factory, _validator
                 )
             self.system_settings = \
-                SystemSettings_v2_3_7_6(
+                SystemSettings_v2_3_7_5(
                     self._session, object_factory, _validator
                 )
             self.tag = \
-                Tag_v2_3_7_6(
+                Tag_v2_3_7_5(
                     self._session, object_factory, _validator
                 )
             self.task = \
-                Task_v2_3_7_6(
+                Task_v2_3_7_5(
                     self._session, object_factory, _validator
                 )
             self.topology = \
-                Topology_v2_3_7_6(
+                Topology_v2_3_7_5(
                     self._session, object_factory, _validator
                 )
             self.user_and_roles = \
-                UserandRoles_v2_3_7_6(
+                UserandRoles_v2_3_7_5(
                     self._session, object_factory, _validator
                 )
             self.users = \
-                Users_v2_3_7_6(
+                Users_v2_3_7_5(
                     self._session, object_factory, _validator
                 )
             self.wireless = \
-                Wireless_v2_3_7_6(
+                Wireless_v2_3_7_5(
                     self._session, object_factory, _validator
                 )
         self.custom_caller = \