+++ test_new.py
@@ -6,7 +6,7 @@ from dnacentersdk import api
 dnac = api.DNACenterAPI(username="devnetuser",
                         password="Cisco123!",
                         base_url="https://sandboxdnac.cisco.com:443",
-                        version='2.3.7.6',
+                        version='2.3.7.5',
                         verify=False)
 
 # Find all devices that have 'Switches and Hubs' in their family