+++ CHANGELOG.md
@@ -9,8 +9,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [2.7.0] - 2024-05-31
 ### Added
-- Add support of DNA Center versions ('2.3.7.6')
-- Adds modules for v2_3_7_6
+- Add support of DNA Center versions ('2.3.7.5')
+- Adds modules for v2_3_7_5
 ### Changed
 - The future library was removed
 - The past library was removed